requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
babel==2.10.1
    # via py-moneyed
black==22.3.0
    # via -r requirements.in
boto3==1.22.0
    # via -r requirements.in
botocore==1.25.0
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   black
    #   geocoder
decorator==5.1.1
    # via ratelim
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.0.4
    # via
    #   -r requirements.in
    #   django-admin-numeric-filter
    #   django-import-export
    #   django-jsonform
    #   django-localflavor
    #   django-money
    #   django-more-admin-filters
    #   django-phonenumber-field
    #   django-storages
django-admin-numeric-filter @ git+https://github.com/vgalaktionov/django-admin-numeric-filter@master
    # via -r requirements.in
django-countries==7.3.2
    # via -r requirements.in
django-import-export==2.8.0
    # via -r requirements.in
django-jsonform==2.9.0
    # via -r requirements.in
django-localflavor==3.1
    # via -r requirements.in
django-money==2.1.1
    # via -r requirements.in
django-more-admin-filters @ git+https://github.com/vgalaktionov/django-more-admin-filters@master
    # via -r requirements.in
django-phonenumber-field==6.1.0
    # via -r requirements.in
django-storages==1.12.3
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
flake8==4.0.1
    # via -r requirements.in
future==0.18.2
    # via geocoder
geocoder==1.38.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
markuppy==1.14
    # via tablib
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
pathspec==0.9.0
    # via black
phonenumbers==8.12.47
    # via -r requirements.in
platformdirs==2.5.2
    # via black
psycopg2-binary==2.9.3
    # via -r requirements.in
py-moneyed==1.2
    # via django-money
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
python-dateutil==2.8.2
    # via botocore
python-stdnum==1.17
    # via django-localflavor
pytz==2022.1
    # via babel
pyyaml==6.0
    # via tablib
ratelim==0.1.6
    # via geocoder
requests==2.27.1
    # via geocoder
s3transfer==0.5.2
    # via boto3
six==1.16.0
    # via
    #   geocoder
    #   python-dateutil
    #   transliterate
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
tomli==2.0.1
    # via black
transliterate==1.10.2
    # via -r requirements.in
typing-extensions==4.2.0
    # via
    #   black
    #   django-countries
urllib3==1.26.9
    # via
    #   botocore
    #   requests
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
